Move TaskGroup to util package
